support VARIANT for pypaimon
